fix: fix pagination component
